feat : deleteMember 적용, security 적용

+++ src/main/java/com/hh99/lv4/domain/comment/service/CommentService.java
@@ -7,7 +7,12 @@ import com.hh99.lv4.domain.comment.entity.Comment;
 import com.hh99.lv4.domain.comment.repository.CommentRepository;
 import com.hh99.lv4.domain.lecture.entity.Lecture;
 import com.hh99.lv4.domain.lecture.service.LectureService;
+import com.hh99.lv4.domain.member.entity.Member;
+import com.hh99.lv4.domain.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,11 +25,17 @@ public class CommentService {
 
     private final CommentRepository commentRepository;
     private final LectureService lectureService;
+    private final MemberService memberService;
 
     public CommentResponseDto createComment(long lectureId, CommentPostDto postDto) {
         Lecture lecture = lectureService.findLectureById(lectureId);
+
+        UserDetails loginMember = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String email = loginMember.getUsername();
+        Member foundMember = memberService.findMemberByEmail(email);
+
         Comment comment = Comment.builder()
-//                .member(memberId)
+                .member(foundMember)
                 .lecture(lecture)
                 .content(postDto.getContent())
                 .build();
@@ -36,9 +47,14 @@ public class CommentService {
     public CommentResponseDto createReply(long lectureId, long parentCommentId, CommentPostDto postDto) {
         Comment parentComment = findCommentById(parentCommentId);
         Lecture lecture = lectureService.findLectureById(lectureId);
+
+        UserDetails loginMember = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String email = loginMember.getUsername();
+        Member foundMember = memberService.findMemberByEmail(email);
+
         Comment reply = Comment.builder()
                 .lecture(lecture)
-                .member(parentComment.getMember())
+                .member(foundMember)
                 .parentCommentId(parentCommentId)
                 .content(postDto.getContent())
                 .build();
@@ -49,15 +65,20 @@ public class CommentService {
 
     public CommentResponseDto updateComment(long commentId, CommentPatchDto patchDto) {
         Comment comment = findCommentById(commentId);
-        //security memberDetails로 유저 검증 예정
-        comment.updateComment(patchDto.getContent());
+        Member member = (Member) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (member.getMemberId() == comment.getMember().getMemberId()) {
+            comment.updateComment(patchDto.getContent());
+        } else throw new SecurityException("댓글 작성자만 수정 할 수 있습니다.");
+
         return CommentResponseDto.fromEntity(comment);
     }
 
     public void deleteComment(long commentId) {
         Comment comment = findCommentById(commentId);
-        //security memberDetails로 유저 검증 예정
-        commentRepository.deleteById(commentId);
+        Member member = (Member) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (member.getMemberId() == comment.getMember().getMemberId()) {
+            commentRepository.deleteById(commentId);
+        } else throw new SecurityException("댓글 작성자만 삭제 할 수 있습니다.");
     }
 
     public Comment findCommentById(long commentId) {
